Changed seed name from 'newSeed' to 'rawMongoSeed';
